+++ tests/test_config_parser.py
@@ -42,13 +42,13 @@ def test_environment_variable_config(capsys: pytest.CaptureFixture[str]) -> None
         "DISABLE_COMMAND_VERIFICATION",
     ]
     expected_device_string = (
-        "address=MSO54-123456,connection_type=TCPIP,device_type=SCOPE,lan_device_name=hislip0"
+        "address=MSO54-123456,connection_type=TCPIP,device_type=SCOPE,lan_device_endpoint=hislip0"
     )
     expected_device = DeviceConfigEntry(
         address="MSO54-123456",
         connection_type=ConnectionTypes.TCPIP,
         device_type=DeviceTypes.SCOPE,
-        lan_device_name="hislip0",
+        lan_device_endpoint="hislip0",
     )
     expected_entry = MappingProxyType({"SCOPE 1": expected_device})
 
@@ -238,7 +238,7 @@ def test_file_config_non_default_path(
             address="MSO54-123456",
             connection_type=ConnectionTypes.TCPIP,
             device_type=DeviceTypes.SCOPE,
-            lan_device_name="hislip0",
+            lan_device_endpoint="hislip0",
         ),
         "SCOPE 2": DeviceConfigEntry(
             address="789.56.4.123",